refactor: split background/index.ts into focused modules
Extract 3 modules from background/index.ts (561 → 93 lines):
- validation.ts: validateSender, validateMessageContent, validateNoteData
- obsidian-handlers.ts: createObsidianClient, isClientError, handleSave,
  handleGetFile, handleTestConnection
- output-handlers.ts: handleMultiOutput and internal helpers
  (offscreen lifecycle, download, clipboard, executeOutput)

index.ts retains: imports, migrateSettings, onMessage listener, handleMessage router

Pure refactoring — no behavioral changes. All 625 tests pass.
